pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cineinfini-audit"
version = "0.3.0"
description = "Video quality auditor for T2V generation"
authors = [{name="CineInfini", email="cineinfini@example.com"}]
license = {text="MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21", "opencv-python>=4.5", "scipy>=1.7",
    "scikit-image>=0.19", "torch>=1.12", "torchvision>=0.13",
    "pillow>=9.0", "tqdm>=4.64", "pydantic>=1.10",
    "scenedetect>=0.5.5", "clip@git+https://github.com/openai/CLIP.git"
]
[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "black", "isort", "flake8"]
dtw = ["dtw-python>=1.3"]
[project.scripts]
cineinfini-audit = "cineinfini.cli:main"

[tool.pytest.ini_options]
markers = ["integration: marks tests that download large files", "slow: slow tests"]
addopts = "-v --tb=short"
